+++ src/pyFAI/detectors/_rayonix.py
@@ -83,9 +83,9 @@ class _Rayonix(Detector):
         :type bin_size: int or (int, int)
         """
         if "__len__" in dir(bin_size) and len(bin_size) >= 2:
-            bin_size = int(round(float(bin_size[0]))), int(round(float(bin_size[1])))
+            bin_size = round(float(bin_size[0])), round(float(bin_size[1]))
         else:
-            b = int(round(float(bin_size)))
+            b = round(float(bin_size))
             bin_size = (b, b)
         if bin_size != self._binning:
             if (bin_size[0] in self.BINNED_PIXEL_SIZE) and (bin_size[1] in self.BINNED_PIXEL_SIZE):
