+++ src/pyFAI/opencl/test/test_collective.py
@@ -103,7 +103,7 @@ class TestGroupFunction(unittest.TestCase):
         tests the sum_int_reduction function
         """
         # rec_workgroup = self.program.test_sum_int_reduction.get_work_group_info(pyopencl.kernel_work_group_info.WORK_GROUP_SIZE, self.ctx.devices[0])
-        maxi = int(round(numpy.log2(min(self.shape,self.max_valid_wg))))+1
+        maxi = round(numpy.log2(min(self.shape,self.max_valid_wg)))+1
         for i in range(maxi):
             wg = 1 << i
             try:
@@ -128,7 +128,7 @@ class TestGroupFunction(unittest.TestCase):
         tests the sum_int_atomic function
         """
 
-        maxi = int(round(numpy.log2(min(self.shape, self.max_valid_wg))))+1
+        maxi = round(numpy.log2(min(self.shape, self.max_valid_wg)))+1
         for i in range(maxi):
             wg = 1 << i
             try:
@@ -154,7 +154,7 @@ class TestGroupFunction(unittest.TestCase):
         """
         data_d = pyopencl.array.to_device(self.queue, self.data.astype("float32"))
         scan_d = pyopencl.array.empty_like(data_d)
-        maxi = int(round(numpy.log2(min(self.shape, self.max_valid_wg))))+1
+        maxi = round(numpy.log2(min(self.shape, self.max_valid_wg)))+1
         for i in range(maxi):
             wg = 1 << i
             try:
@@ -180,7 +180,7 @@ class TestGroupFunction(unittest.TestCase):
         """
         data_d = pyopencl.array.to_device(self.queue, self.data.astype("float32"))
         scan_d = pyopencl.array.empty_like(data_d)
-        maxi = int(round(numpy.log2(min(self.shape/2, self.max_valid_wg))))+1
+        maxi = round(numpy.log2(min(self.shape/2, self.max_valid_wg)))+1
         for i in range(maxi):
             wg = 1 << i
             try:
@@ -209,7 +209,7 @@ class TestGroupFunction(unittest.TestCase):
         """
         data_d = pyopencl.array.to_device(self.queue, self.data.astype("float32"))
         scan_d = pyopencl.array.empty_like(data_d)
-        maxi = int(round(numpy.log2(min(self.shape/2, self.max_valid_wg))))+1
+        maxi = round(numpy.log2(min(self.shape/2, self.max_valid_wg)))+1
         for i in range(maxi):
             wg = 1 << i
             try:
@@ -242,7 +242,7 @@ class TestGroupFunction(unittest.TestCase):
         numpy.random.shuffle(data)
         data_d = pyopencl.array.to_device(self.queue, data)
 
-        maxi = int(round(numpy.log2(self.shape)))+1
+        maxi = round(numpy.log2(self.shape))+1
         for i in range(5,maxi):
             wg = 1 << i
 
@@ -282,7 +282,7 @@ class TestGroupFunction(unittest.TestCase):
         numpy.random.shuffle(data)
         data_d = pyopencl.array.to_device(self.queue, data)
 
-        maxi = int(round(numpy.log2(self.shape)))+1
+        maxi = round(numpy.log2(self.shape))+1
         for i in range(5,maxi):
             wg = 1 << i
